+++ arrow-ord/src/cmp.rs
@@ -601,8 +601,8 @@ impl<'a, T: ByteViewType> ArrayOrd for &'a GenericByteViewArray<T> {
 
             // The lower 32 bits encode the length (little-endian),
             // the upper 96 bits hold the actual data
-            let l_len = l_bits as u32 as u64;
-            let r_len = r_bits as u32 as u64;
+            let l_len = l_bits as u32;
+            let r_len = r_bits as u32;
 
             // Remove the length bits, leaving only the data
             let l_data = l_bits >> 32;
@@ -679,8 +679,8 @@ pub fn compare_byte_view<T: ByteViewType>(
 
         // The lower 32 bits encode the length (little-endian),
         // the upper 96 bits hold the actual data
-        let l_len = l_bits as u32 as u64;
-        let r_len = r_bits as u32 as u64;
+        let l_len = l_bits as u32;
+        let r_len = r_bits as u32;
 
         // Remove the length bits, leaving only the data
         let l_data = l_bits >> 32;
